Update README.md
Changed the need to look at the config.php file used for manual installs on hosted sites. Also, the location of the /css/styles.css file and the /img/ directory may need to be configured manually to fit the location of the root directory. I have ../img/ as well ../css/styles.css but that path may not meet your need. Make those changes in your individual files.